Swapped some instances of "Hi" for "Hello"
To make it clear that it's "you" that is breaching the etiquette, not the co-worker, I have swapped the initial greetings so it's "you" that says "hello", tying it back to the name of the article. I know that the point is not about "hello" vs "hi". But first time reading the page this momentarily confused me and distracted me from the point.

I have changed a few more "Hi"s to "Hello"s for consistency but allowed the interchangeability of "Hello" and "Hi" to come through by the end of the article, so the reader is not left with the impression that "Hello" vs "Hi" is the point.